2025 한이음/논문 관련/음성 생성기/키오스크 관련_voice_maker/requirements.txt
openai>=1.40.0
pandas
python-dotenv
tqdm
